Hide hiring message on landing page
