+++ src/components/SectionContent.tsx
@@ -15,7 +15,6 @@ const Container = styled("div")({
   marginRight: `${SPACING.default}%`,
   maxWidth: `${46 * SPACING.default}px`,
   lineHeight: 1.4,
-  overflow: "hidden",
 });
 
 export default SectionContent;
